requirements.txt
# =============================================================================
# Genotek Sales Support Telegram Bot -- Dependencies
# =============================================================================

# Telegram
python-telegram-bot>=21.0

# AWS Bedrock
boto3>=1.34.0

# Anthropic Direct API (uncomment if USE_BEDROCK=false)
# anthropic>=0.39.0

# LangChain + LangGraph
langchain-core>=0.3.0
langgraph>=0.2.0

# Supabase
supabase>=2.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
